# openapi.json
{"openapi":"3.1.0","info":{"title":"企业风险评估系统","version":"0.1.0"},"paths":{"/api/v1/enterprise/pk":{"get":{"tags":["enterprise"],"summary":"Enterprise Pk","operationId":"enterprise_pk_api_v1_enterprise_pk_get","parameters":[{"name":"ids","in":"query","required":true,"schema":{"type":"string","description":"逗号分隔的企业ID，如 ENT001,ENT002","title":"Ids"},"description":"逗号分隔的企业ID，如 ENT001,ENT002"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/enterprise/{enterprise_id}":{"get":{"tags":["enterprise"],"summary":"Get Enterprise","operationId":"get_enterprise_api_v1_enterprise__enterprise_id__get","parameters":[{"name":"enterprise_id","in":"path","required":true,"schema":{"type":"string","title":"Enterprise Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/enterprise/{enterprise_id}/dimensions":{"get":{"tags":["enterprise"],"summary":"Get Enterprise Dimensions","operationId":"get_enterprise_dimensions_api_v1_enterprise__enterprise_id__dimensions_get","parameters":[{"name":"enterprise_id","in":"path","required":true,"schema":{"type":"string","title":"Enterprise Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/risk/warnings":{"get":{"tags":["risk"],"summary":"List Warnings","operationId":"list_warnings_api_v1_risk_warnings_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/health":{"get":{"summary":"Health","operationId":"health_api_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}